pyproject.toml
[tool.poetry]
name = "ult-edu"
version = "0.1.0"
description = ""
authors = ["Sergei Razin <900556@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
djangorestframework-xml = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
